rag-api/pyproject.toml
[project]
name = "rag-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",
    "pydantic-settings>=2.6.0",
    "msgspec>=0.18.0",
    "alembic>=1.14.0",
    "python-multipart>=0.0.12",
    "anthropic>=0.40.0",
    "sentence-transformers>=3.0.0",
    "pypdf>=5.0.0",
    "pyjwt>=2.11.0",
    "boto3>=1.42.63",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-psycopg2>=2.9.0",
    "moto[s3]>=5.1.22",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
